Cleanup

+++ sunray/helper.cpp
@@ -131,8 +131,8 @@ float distance(float x1, float y1, float x2, float y2){
   return sqrtf( (x1-x2)*(x1-x2) + (y1-y2)*(y1-y2) );
 }
 
-int sign(double x) { 
-  return x<0 ? -1 : 1; 
+float sign(float x) { 
+  return x < 0.0 ? -1.0 : 1.0; 
 }
 
 double deg2rad(double deg) {
